+++ src/imio/actionspanel/browser/transitions.py
@@ -30,8 +30,8 @@ class ConfirmTransitionView(BrowserView):
             # while the Cancel button is hit
             self.request.response.redirect(form.get('form.HTTP_REFERER'))
         elif submitted:
-            uid_catalog = getToolByName(self.context, 'uid_catalog')
-            obj = uid_catalog(UID=self.request['objectUid'])[0].getObject()
+            portal_catalog = getToolByName(self.context, 'portal_catalog')
+            obj = portal_catalog(UID=self.request['objectUid'])[0].getObject()
             return obj.restrictedTraverse('@@actions_panel').triggerTransition(self.request.get('transition'),
                                                                                self.request.get('comment'))
         return self.index()